Create index.md
This is my first page hosted on github!

This sketch started out just trying to make a spiraling image by rotating a square while reducing its side length and also changing its color. Then I decided I wanted the spiral to display one square at a time...and that was the puzzle for me. I tried using setInterval, making my main function an IIFE, and using async/await, but didn't really know enough to make them work yet. But then I read about using frameCount, and that worked to slow down my display loop, and by displaying as long as the franeCount was AT LEAST a certain number, the squares persisted on the canvas. 

It took me some googling and head-scratching to make this work and I was very pleased when it finally did!